fix: resolve Next.js 15 cookies() Promise issue and implement admin auth fallback
- Fix TypeScript error in admin layout by awaiting cookies() function
- Replace old layout.tsx with fixed version that handles Next.js 15 properly
- Implement temporary admin authentication system for emergency access
- Add fallback authentication when Kinde session fails
- Ensure admin layout compiles successfully with Next.js 15
